Move resource package to ui module.
