+++ mvp_example/build.gradle
@@ -1,6 +1,7 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-kapt'
 apply from: 'config.gradle'
 
 android {
@@ -52,6 +53,7 @@ dependencies {
 
     implementation project(path: ':domain')
     implementation project(path: ':mvp')
+    implementation project(path: ':ui')
 
     // Platform libraries
     implementation libraries.platform.kotlinStdLib
