+++ settings.gradle
@@ -7,3 +7,4 @@ include ':mvi_example'
 include ':mvp_example'
 include ':mvvm_example'
 include ':domain'
+include ':ui'
